깊은복사설명, delete fail alert 창 띄우기 finish

+++ lib/views/product/list/product_list_view.dart
@@ -1,11 +1,13 @@
 import 'package:data_app/controller/product_controller.dart';
 import 'package:data_app/domain/product/product.dart';
+import 'package:data_app/views/components/my_alert_dialog.dart';
 import 'package:data_app/views/product/list/product_list_view_store.dart';
+import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
-class ProductListPage extends ConsumerWidget {
-  const ProductListPage({Key? key}) : super(key: key);
+class ProductListView extends ConsumerWidget {
+  const ProductListView({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context, WidgetRef ref) {
@@ -26,7 +28,11 @@ class ProductListPage extends ConsumerWidget {
         itemBuilder: (context, index) => ListTile(
           key: ValueKey(pm[index].id),
           onTap: (){
-            pc.deleteById(pm[index].id);
+            if(pm[index].id != 4){
+              pc.deleteById(pm[index].id);
+            }else{
+              showCupertinoDialog(context: context, builder: (context) => MyAlertDialog(msg: "삭제실패"));
+            }
           },
           leading: Icon(Icons.account_balance_wallet),
           title: Text("${pm[index].name}", style: TextStyle(fontWeight: FontWeight.bold)),
